Fixing bugs found in endpoints and adding more

+++ backend/src/services/tasks.ts
@@ -1,12 +1,11 @@
 import { db } from "../db";
-import { TaskTypes } from "../generated/prisma/enums";
+import { Status, TaskTypes } from "../generated/prisma/enums";
 import { JsonObject } from "../generated/prisma/internal/prismaNamespace";
-
+import { TaskAssignment } from "../generated/prisma/client";
 
 async function getAllTasks() {
     return await db.task.findMany();
 }
-
 
 async function getTaskByID(taskID: Number) {
     return await db.task.findUnique({
@@ -16,7 +15,6 @@ async function getTaskByID(taskID: Number) {
     });
 }
 
-
 async function getAllTasksByUserID(userID: Number) {
     return await db.task.findMany({
         where: {
@@ -25,17 +23,30 @@ async function getAllTasksByUserID(userID: Number) {
     });
 }
 
+async function getTaskAllTaskAssignments() {
+    return await db.taskAssignment.findMany();
+}
+
+async function getTaskAssignmentsByUserID(userID: Number) {
+    return await db.taskAssignment.findMany({
+        where: {
+            assigneeID: Number(userID),
+        },
+    });
+}
+
+async function getTaskAssignmentsByTaskID(taskID: Number) {
+    return await db.taskAssignment.findMany({
+        where: {
+            taskID: Number(taskID),
+        },
+    });
+}
 
 async function createTask(publisherID: Number, body: JsonObject) {
-    let { 
-        name, 
-        type, 
-        payment, 
-        dueDate, 
-        description, 
-        images, 
-        hashtags 
-    } = body;
+    let { name, type, payment, dueDate, description, images, hashtags } = body;
+
+    //TODO chedk if task is unique
 
     return await db.task.create({
         data: {
@@ -58,13 +69,25 @@ async function createTask(publisherID: Number, body: JsonObject) {
     });
 }
 
-
 async function assignTask(taskID: Number, userID: Number) {
     const task = await db.task.findUnique({
         where: { taskID: Number(taskID) },
         select: { publisherID: true },
     });
     if (!task) throw new Error(`Task ${taskID} not found`);
+
+    const user = await db.taskAssignment.findFirst({
+        where: {
+            taskID: Number(taskID),
+            assigneeID: Number(userID),
+        },
+    });
+
+    if (user)
+        throw new Error(`User ${userID} is already assigned to task ${taskID}`);
+
+    if (task.publisherID === Number(userID))
+        throw new Error("Publisher cannot be assigned to their own task");
 
     await db.$transaction([
         //EITHER ALL SUCCEED OR ALL FAIL.
@@ -84,7 +107,6 @@ async function assignTask(taskID: Number, userID: Number) {
     ]);
 }
 
-
 async function unAssignTask(taskID: Number, userID: Number) {
     const conversation = await db.conversation.findFirst({
         //This finds the first Conversation row where both taskCID and userID match. findUnique couldn't be used here because the filter is on a combination of non-unique fields — findUnique requires a single unique/primary key field.
@@ -99,6 +121,25 @@ async function unAssignTask(taskID: Number, userID: Number) {
             `Conversation for task ${taskID} and user ${userID} not found`
         );
 
+    const user = await db.taskAssignment.findFirst({
+        where: {
+            taskID: Number(taskID),
+            assigneeID: Number(userID),
+        },
+    });
+
+    if (!user)
+        throw new Error(`User ${userID} is not assigned to task ${taskID}`);
+
+    const task = await db.task.findUnique({
+        where: { taskID: Number(taskID) },
+        select: { publisherID: true },
+    });
+    if (!task) throw new Error(`Task ${taskID} not found`);
+
+    if (task.publisherID === Number(userID))
+        throw new Error("Publisher can't be unassigned from their own task.");
+
     await db.$transaction([
         db.taskAssignment.deleteMany({
             where: {
@@ -112,8 +153,14 @@ async function unAssignTask(taskID: Number, userID: Number) {
     ]);
 }
 
-
 async function updateTask(taskID: Number, body: JsonObject) {
+    const task = await getTaskByID(taskID);
+
+    if (!task)
+        throw new Error(
+            `Task ${taskID} can't be updated as it does not exist.`
+        );
+
     let {
         name,
         type,
@@ -131,7 +178,7 @@ async function updateTask(taskID: Number, body: JsonObject) {
         data: {
             name: name as string | undefined,
             type: type as TaskTypes | undefined,
-            status: status as any | undefined,
+            status: status as Status | undefined,
             payment: payment as number | undefined,
             dueDate: dueDate ? new Date(dueDate as string) : undefined,
             completedDate: completedDate
@@ -152,7 +199,6 @@ async function updateTask(taskID: Number, body: JsonObject) {
     });
 }
 
-
 async function deleteTask(taskID: Number) {
     await db.task.delete({
         where: {
@@ -163,11 +209,13 @@ async function deleteTask(taskID: Number) {
     return;
 }
 
-
 export const taskServices = {
     getAllTasks,
     getTaskByID,
     getAllTasksByUserID,
+    getTaskAllTaskAssignments,
+    getTaskAssignmentsByTaskID,
+    getTaskAssignmentsByUserID,
     assignTask,
     unAssignTask,
     createTask,
